+++ ExpandWorldData/data/DataLoading.cs
@@ -40,6 +40,13 @@ public class DataLoading
     foreach (var item in data)
       ZDOData.Register(item);
   }
+  public static void Save(ZDOData data)
+  {
+    if (Helper.IsClient()) return;
+    var yaml = File.Exists(FilePath) ? File.ReadAllText(FilePath) : DataManager.Serializer().Serialize(DefaultData.Data);
+    yaml += "\n" + DataManager.Serializer().Serialize(new DataData[] { ToData(data) });
+    File.WriteAllText(FilePath, yaml);
+  }
   ///<summary>Loads all yaml files returning the deserialized vegetation entries.</summary>
   private static List<ZDOData> FromFile()
   {
@@ -66,78 +73,79 @@ public class DataLoading
     foreach (var value in data.floats ?? new string[0])
     {
       var split = Parse.Split(value);
-      if (split.Length != 2) continue;
-      zdo.Floats.Add(split[0].GetStableHashCode(), Parse.Float(split[1]));
+      if (split.Length < 2) continue;
+      var str = string.Join(",", split.Skip(1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Floats.Add(hash, Parse.Float(str));
     }
     foreach (var value in data.ints ?? new string[0])
     {
       var split = Parse.Split(value);
-      if (split.Length != 2) continue;
-      zdo.Ints.Add(split[0].GetStableHashCode(), Parse.Int(split[1]));
+      if (split.Length < 2) continue;
+      var str = string.Join(",", split.Skip(1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Ints.Add(hash, Parse.Int(str));
     }
     foreach (var value in data.longs ?? new string[0])
     {
       var split = Parse.Split(value);
-      if (split.Length != 2) continue;
-      zdo.Longs.Add(split[0].GetStableHashCode(), Parse.Long(split[1]));
+      if (split.Length < 2) continue;
+      var str = string.Join(",", split.Skip(1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Longs.Add(hash, Parse.Long(str));
     }
     foreach (var value in data.strings ?? new string[0])
     {
       var split = Parse.Split(value);
-      if (split.Length != 2) continue;
-      zdo.Strings.Add(split[0].GetStableHashCode(), split[1]);
+      if (split.Length < 2) continue;
+      var str = string.Join(",", split.Skip(1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Strings.Add(hash, str);
     }
     foreach (var value in data.vecs ?? new string[0])
     {
       var split = Parse.Split(value);
       if (split.Length != 4) continue;
-      zdo.Vecs.Add(split[0].GetStableHashCode(), Parse.VectorXZY(split, 1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Vecs.Add(hash, Parse.VectorXZY(split, 1));
     }
     foreach (var value in data.quats ?? new string[0])
     {
       var split = Parse.Split(value);
       if (split.Length != 4) continue;
-      zdo.Quats.Add(split[0].GetStableHashCode(), Parse.AngleYXZ(split, 1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.Quats.Add(hash, Parse.AngleYXZ(split, 1));
+    }
+    foreach (var value in data.bytes ?? new string[0])
+    {
+      var split = Parse.Split(value);
+      if (split.Length < 2) continue;
+      var str = string.Join(",", split.Skip(1));
+      var hash = int.TryParse(split[0], out var h) ? h : split[0].GetStableHashCode();
+      zdo.ByteArrays.Add(hash, Convert.FromBase64String(str));
     }
     return zdo;
   }
-  public static VegetationData ToData(ZoneSystem.ZoneVegetation veg)
+  public static DataData ToData(ZDOData zdo)
   {
-    VegetationData data = new()
+    DataData data = new()
     {
-      enabled = veg.m_enable,
-      prefab = veg.m_prefab.name,
-      min = veg.m_min,
-      max = veg.m_max,
-      forcePlacement = veg.m_forcePlacement,
-      scaleMin = veg.m_scaleMin.ToString(CultureInfo.InvariantCulture),
-      scaleMax = veg.m_scaleMax.ToString(CultureInfo.InvariantCulture),
-      randTilt = veg.m_randTilt,
-      chanceToUseGroundTilt = veg.m_chanceToUseGroundTilt,
-      biome = DataManager.FromBiomes(veg.m_biome),
-      biomeArea = DataManager.FromBiomeAreas(veg.m_biomeArea),
-      blockCheck = veg.m_blockCheck,
-      minAltitude = veg.m_minAltitude,
-      maxAltitude = veg.m_maxAltitude,
-      minOceanDepth = veg.m_minOceanDepth,
-      maxOceanDepth = veg.m_maxOceanDepth,
-      minVegetation = veg.m_minVegetation,
-      maxVegetation = veg.m_maxVegetation,
-      minTilt = veg.m_minTilt,
-      maxTilt = veg.m_maxTilt,
-      terrainDeltaRadius = veg.m_terrainDeltaRadius,
-      maxTerrainDelta = veg.m_maxTerrainDelta,
-      minTerrainDelta = veg.m_minTerrainDelta,
-      snapToWater = veg.m_snapToWater,
-      snapToStaticSolid = veg.m_snapToStaticSolid,
-      groundOffset = veg.m_groundOffset,
-      groupSizeMin = veg.m_groupSizeMin,
-      groupSizeMax = veg.m_groupSizeMax,
-      groupRadius = veg.m_groupRadius,
-      inForest = veg.m_inForest,
-      forestTresholdMin = veg.m_forestTresholdMin,
-      forestTresholdMax = veg.m_forestTresholdMax
+      name = zdo.Name,
+      floats = zdo.Floats.Select(pair => $"{DefaultData.Convert(pair.Key)}, {Helper.Print(pair.Value)}").ToArray(),
+      ints = zdo.Ints.Select(pair => $"{DefaultData.Convert(pair.Key)}, {pair.Value}").ToArray(),
+      longs = zdo.Longs.Select(pair => $"{DefaultData.Convert(pair.Key)}, {pair.Value}").ToArray(),
+      strings = zdo.Strings.Select(pair => $"{DefaultData.Convert(pair.Key)}, {pair.Value}").ToArray(),
+      vecs = zdo.Vecs.Select(pair => $"{DefaultData.Convert(pair.Key)}, {Helper.Print(pair.Value)}").ToArray(),
+      quats = zdo.Quats.Select(pair => $"{DefaultData.Convert(pair.Key)}, {Helper.Print(pair.Value)}").ToArray(),
+      bytes = zdo.ByteArrays.Select(pair => $"{DefaultData.Convert(pair.Key)}, {Convert.ToBase64String(pair.Value)}").ToArray(),
     };
+    if (data.floats.Length == 0) data.floats = null;
+    if (data.ints.Length == 0) data.ints = null;
+    if (data.longs.Length == 0) data.longs = null;
+    if (data.strings.Length == 0) data.strings = null;
+    if (data.vecs.Length == 0) data.vecs = null;
+    if (data.quats.Length == 0) data.quats = null;
+    if (data.bytes.Length == 0) data.bytes = null;
     return data;
   }